Fixed audio and made loading screen better
Audio will still require a proper refactor for it to work as intended, options and instructions left as comments.
In built versions, the issue with audio not playing is that for some reason, runtime generated paths don't work as intended, so static path loading is a must.

For the loading screen, there are multiple options, but relevant are the currently existing 3 modes of operation -> Call from LoadScene, which will make it load and transition to a new scene(useful to change from menu to game, from game to credits, cutscenes, etc), call from LoadScene with a target node (useful for loading existing levels for instance), and working on the UI node LoadScereen itself, as just a pop-up.